scripts/rag-indexer/requirements.txt
anthropic>=0.39.0
qdrant-client>=1.12.0
FlagEmbedding>=1.2.10
torch>=2.4.0
langchain-text-splitters>=0.3.2
tiktoken>=0.8.0
python-dotenv>=1.0.1
tqdm>=4.66.0